Run detection phase sequentially per CMP. (#719)
As most rules at this phase are sync, this has no performance cost, but
allows us to reduce the total number of steps to run, because for detections
with multiple steps, we don't have to run subsequent steps if the first
fails.

In the worst case (no CMP on page), reduces total number of check steps
from 4515 to 4326 (4% reduction).